js/xpconnect/tests/chrome/test_bug571849.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500931
-->
<window title="Mozilla Bug 500931"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=500931"
     target="_blank">Mozilla Bug 500931</a>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  /** Test for Bug 500931 **/

  function go() {
    var ifr = document.getElementById('ifr');
    var docnodes = ifr.contentDocument.body.childNodes;
    var output;
    for (let i in Iterator(docnodes))
        output = i;
    is(output[0], "0", "enumerated the 0th element");
    ok(/Text/.test(output[1]), "the 0th element was a text node");
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();

  ]]></script>
  <iframe type="content"
          src="http://example.org/tests/js/xpconnect/tests/mochitest/bug571849_helper.html"
          onload="go()"
          id="ifr">
  </iframe>
  </body>
</window>
